+++ CHANGELOG.md
@@ -7,10 +7,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [vNext] Unreleased
 
+## [0.0.3] 2018-01-30
+
+### Added
+
+- Compare-ScriptOutput.ps1 implemented function process() part to fill up excel file.
+- Private function Get-ExcelColumnInt convert-an-excel-column-letter-into-its-number.
+
+## [0.0.2] 2018-01-30
+
+### Added
+
+- Compare-ScriptOutput.ps1 implemented function begin() part to Vester result to a variable.
+
+### Changed
+
+- VMwareBaselineCheck.psm1 changed public and private folder
+
 ### Fixed
 
-- test file
+- Tests file
 
 ## [0.0.1] 2018-01-20
 
-initial release
+- initial release with scaffold